move models from graphql file to separated directory
